+++ octopus/src/main/kotlin/com/koupper/octopus/logging/GlobalLogger.kt
@@ -1,0 +1,13 @@
+package com.koupper.octopus.logging
+
+object GlobalLogger {
+    @Volatile
+    private var delegate: Logger? = null
+
+    fun setLogger(logger: Logger) {
+        delegate = logger
+    }
+
+    val log: Logger
+        get() = delegate ?: LoggerFactory.get("GlobalLogger")
+}